+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Session timezone support. Refs UID-204.
 * Tolerate missing `name` attribute when sorting module descriptors. Refs UID-200.
 * Remove display of mod-configuration values. Refs UID-182.
+* Display mod-scheduler's timers. Refs UID-208.
 
 ## [10.0.0](https://github.com/folio-org/ui-developer/tree/v10.0.0) (2025-03-17)
 [Full Changelog](https://github.com/folio-org/ui-developer/compare/v9.0.0...v10.0.0)